Várias mudanças e várias atualizações

+++ src/main/kotlin/net/perfectdreams/pantufa/tables/Profiles.kt
@@ -1,5 +1,5 @@
 package net.perfectdreams.pantufa.tables
 
 object Profiles : SnowflakeTable() {
-	val money = double("money").index()
+	val money = long("money").index()
 }
